django-api/pyproject.toml
[tool.poetry]
name = "mainframe"
version = "0.1.0"
description = ""
authors = ["Andrei Prădan <andrei@pradan.dev>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.1.5"
django-cors-headers = "^3.13.0"
django-environ = "0.9.0"
djangorestframework = "3.14.0"
gunicorn = "20.1.0"
psycopg2-binary = "2.9.5"
pyjwt = "2.6.0"
python-telegram-bot = "13.11"
requests = "^2.28.1"
python-dotenv = "^0.21.0"
pygithub = "^1.57"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
set-hooks = 'deploy.hooks:set_hooks'
full-setup = 'deploy.setup:full_setup'
